철학자 시간에 대해서 사소한 부분 고쳐보았음

+++ 10.philosophers/philo-one/philo.h
@@ -43,7 +43,7 @@ typedef struct s_philo
 {
 	pthread_t thread;
 	int whoami;
-	int when_eat;
+	unsigned long when_eat;
 	int left_fork_num;
 	int right_fork_num;
 	int meal_num;
